Added new styling, randomize of questions and calculations to the results view.

+++ Domain/Models/Result.cs
@@ -3,9 +3,9 @@
     public class Result
     {
         public string Question { get; set; }
-        public int RespondentsWeight { get; set; }
-        public int ExpectationGap { get; set; }
-        public int Accuracy { get; set; }
-        public int ManagersWeight { get; set; }
+        public decimal RespondentsWeight { get; set; }
+        public decimal ExpectationGap { get; set; }
+        public decimal Accuracy { get; set; }
+        public decimal ManagersWeight { get; set; }
     }
 }
